requirements.txt
AdvancedHTTPServer==2.2.0
aiodns==2.0.0
aiohttp==3.6.2
alembic==1.1.0.dev0
amqp==1.4.9
aniso8601==8.0.0
anyjson==0.3.3
asgiref==3.2.7
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
Babel==2.8.0
backcall==0.1.0
backdoor-factory==0.0.0
basemap==1.2.1
bcrypt==3.1.7
beautifulsoup4==4.9.0
billiard==3.3.0.23
binwalk==2.2.0
blinker==1.4
boltons==19.1.0
bottle==0.12.15
Brlapi==0.7.0
Brotli==1.0.7
cairocffi==0.9.0
capstone==3.0.5
cbor==1.0.0
celery==3.1.18
certifi==2019.11.28
cffi==1.14.0
Chameleon==3.6.2
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==6.7
click-plugins==1.1.1
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
crcelk==1.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
debtags==2.1
decorator==4.4.2
deprecation==2.0.7
devscripts==2.20.3
dhcpig==0.0.0
dicttoxml==1.7.4
distro==1.4.0
distro-info==0.23
Django==3.0.7
django-cors-headers==3.3.0
django-extensions==2.2.9
django-nose==1.4.6
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
dnslib==0.9.12
dnspython==1.16.0
drf-util==0.0.12
drf-yasg==1.17.1
ecdsa==0.15
elasticsearch==6.3.1
email-validator==1.0.3
entrypoints==0.3
Faker==4.1.0
faraday-plugins==1.0.2
faradaysec==3.10.2
feedparser==5.2.1
filedepot==0.5.2
filteralchemy==0.1.0
flasgger==0.9.4
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Classful==0.14.1
Flask-KVSession-fork==0.6.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.8
Flask-Restless==0.17.0
Flask-Security==3.0.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
flex==6.13.2
future==0.18.2
GDAL==3.0.4
GeoIP==1.3.2
geoip2==2.9.0
geojson==2.5.0
gevent==1.4.0
gpg===1.13.1-unknown
graphene==2.1.7
graphene-sqlalchemy==2.1.2
graphql-core==2.2.1
graphql-relay==2.0.0
greenlet==0.4.15
grequests==0.4.0
h11==0.9.0
h2==3.2.0
hashID==3.1.4
hpack==3.0.0
html2text==2020.1.16
html5lib==1.0.1
httplib2==0.14.0
hupper==1.10.2
hyperframe==5.2.0
hyperlink==19.0.0
icalendar==4.0.3
idna==2.7
impacket==0.9.20
incremental==16.10.1
inflection==0.5.0
IPy==1.0
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
itypes==1.2.0
jedi==0.16.0
jeepney==0.4.3
Jinja2==2.10.1
jsonpointer==1.14
jsonschema==3.0.2
jupyter-core==4.6.3
kaitaistruct==0.8
kazam==1.4.5
keyring==18.0.1
keyrings.alt==3.4.0
KismetCaptureFreaklabsZigbee==2018.7.0
KismetCaptureRtl433==2019.9.1
KismetCaptureRtladsb==2019.10.1
KismetCaptureRtlamr==2019.10.1
kiwisolver==1.0.1
kombu==3.0.37
ldap3==2.5.1
ldapdomaindump==0.9.1
louis==3.13.0
lxml==4.5.0
lz4==3.0.2+dfsg
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
marshmallow==3.0.0b3
marshmallow-sqlalchemy==0.19.0
matplotlib==3.2.1
maxminddb==1.4.1
mechanize==0.4.5
mimerender==0.6.0
mistune==0.8.4
mitmproxy==4.0.4
msgpack==0.6.2
multidict==4.7.5
mysqlclient==1.4.4
nassl==3.0.0
nbformat==5.0.7
netaddr==0.7.19
nose==1.3.7
nplusone==1.0.0
numpy==1.17.4
olefile==0.46
packaging==20.3
paramiko==2.6.0
parso==0.6.2
passlib==1.7.2
Paste==3.4.0
PasteDeploy==2.1.0
PasteScript==2.0.2
pefile==2019.4.18
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.2.1
plaster==1.0
plaster-pastedeploy==0.5
plecost==1.1.2
plotly==4.6.0
pluginbase==1.0.0
ply==3.11
prettytable==0.7.2
promise==2.2
prompt-toolkit==3.0.5
protobuf==3.11.4
psycopg2==2.8.5
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.11
pycairo==1.16.2
pycares==3.1.1
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.6.1
pycups==1.9.73
pycurl==7.43.0.2
pydot==1.4.1
pyenchant==2.0.0
Pygments==2.3.1
PyGObject==3.36.0
pygtkspellcheck==4.0.5
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymssql==2.1.4
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyotp==2.3.0
pyparsing==2.4.6
pypcapfile==0.12.1
pyperclip==1.8.0
pypng==0.0.20
pyproj==2.6.0
PyQRCode==1.2.1
PyQt5==5.14.2
pyramid==1.10.4
pyrsistent==0.15.5
pyserial==3.4
pyshp==2.1.0
PySimpleSOAP==1.16.2
pysmbc==1.0.15.6
PySocks==1.6.8
python-apt==2.1.3
python-dateutil==2.8.1
python-debian==0.1.37
python-debianbts==3.0.2
python-editor==1.0.3
python-magic==0.4.16
python-magic-ahupp==0.4.13
python-mimeparse==1.6.0
python-pam==1.8.4
python-snappy==0.5.3
PyTrie==0.2
pytz==2019.3
pyxdg==0.26
PyYAML==3.13
qrcode==6.1
redis==2.10.3
reportbug==7.6.0
repoze.lru==0.7
requests==2.20.1
requests-file==1.5.1
retrying==1.3.3
rfc3987==1.3.8
rq==1.2.2
ruamel.yaml==0.15.89
rule-engine==1.1.0
Rx==1.6.1
scapy==2.4.3
scipy==1.3.3
SecretStorage==3.1.2
selenium==4.0.0a1
service-identity==18.1.0
shodan==1.23.0
simplejson==3.17.0
simplekv==0.13.0
sip==4.19.22
six==1.14.0
smoke-zephyr==2.0.0
sortedcontainers==2.1.0
soupsieve==2.0
speaklater==1.3
SQLAlchemy==1.3.15
sqlalchemy-schemadisplay==1.3
sqlparse==0.2.4
sslyze==3.0.3
strict-rfc3339==0.7
swagger-spec-validator==2.4.1
syslog-rfc5424-formatter==1.2.2
tabulate==0.8.2
Tempita==0.5.2
termcolor==1.1.0
termineter==1.0.4
text-unidecode==1.3
texttable==1.6.2
theHarvester==3.1.0
tld==0.11.11
tls-parser==1.2.1
tornado==5.1.1
tqdm==4.43.0
traitlets==4.3.3
translationstring==1.3
Twisted==18.9.0
txaio==2.10.0
typing-extensions==3.7.4.2
tzlocal==2.1b1
u-msgpack-python==2.3.0
unattended-upgrades==0.1
unicodecsv==0.14.1
Unidecode==1.1.1
unidiff==0.5.5
uritemplate==3.0.1
urllib3==1.24.3
urwid==2.0.1
validate-email==1.3
vboxapi==1.0
venusian==3.0.0
vinetto==0.8.0
wafw00f==2.1.0
waitress==1.4.1
wapiti3==3.0.3
wcwidth==0.1.9
webargs==5.1.2
webencodings==0.5.1
WebOb==1.8.6
websocket-client==0.53.0
WebTest==2.0.34
Werkzeug==1.0.0
wfuzz==2.4.5
wifite==2.5.2
wsaccel==0.6.2
wsproto==0.15.0
WTForms==2.2.1
xcffib==0.8.1
XlsxWriter==1.1.2
yara-python==3.10.0
yarl==1.4.2
yaswfp==0.9.3
zim==0.72.1
zope.component==4.3.0
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==4.7.1
